[refactor] Create the SET_COMPARISON_FUNCTIONS once
Instead of creating it at runtime all the time
